break-out action-creators into dedicated file: "src/redux/epics/index.js"
